feat: ввёл базис для process type разделения в проекте
