+++ src/pybot/health/routers/health.py
@@ -1,0 +1,32 @@
+from dishka.integrations.fastapi import DishkaRoute, FromDishka
+from fastapi import APIRouter
+
+from ...dto.health_dto import HealthStatusDTO
+from ...services.health import HealthService
+
+router = APIRouter(route_class=DishkaRoute)
+
+
+@router.get(
+    "/health",
+    response_model=HealthStatusDTO,
+    response_model_exclude_none=True,
+    tags=["health"],
+    summary="Liveness probe",
+    description="Returns OK if the process is running. No external dependencies are checked.",
+    responses={
+        200: {
+            "description": "Process is alive.",
+        },
+    },
+)
+async def health(health_service: FromDishka[HealthService]) -> HealthStatusDTO:
+    """Return liveness status of the process.
+
+    Args:
+        health_service: Health application service resolved from Dishka container.
+
+    Returns:
+        HealthStatusDTO: Liveness status payload.
+    """
+    return await health_service.health()
